feat: use G6 for graph visualization
