+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "cultivator",
   "dependencies": {
+    "@antv/g6": "^4.8.7",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@fontsource/roboto": "^4.5.8",
